Refactor and Whitespace Trim
